# Fix performance of some versions of add_agent